Moved BuildingData scriptable object to separate file
